+++ src/tapps_mcp/server.py
@@ -2,6 +2,11 @@
 
 Creates the FastMCP server instance, registers all tools, and provides
 ``run_server()`` for the CLI.
+
+Tool handlers are split across modules for maintainability:
+  - ``server_scoring_tools``: tapps_score_file, tapps_quality_gate, tapps_quick_check
+  - ``server_pipeline_tools``: tapps_validate_changed, tapps_session_start, tapps_init
+  - ``server_metrics_tools``: tapps_dashboard, tapps_stats, tapps_feedback, tapps_research
 """
 
 from __future__ import annotations
@@ -34,7 +39,7 @@ mcp = FastMCP("TappsMCP")
 
 
 # ---------------------------------------------------------------------------
-# Helpers
+# Helpers (shared by tool modules via lazy import)
 # ---------------------------------------------------------------------------
 
 
@@ -50,12 +55,7 @@ def _normalize_path_for_mapping(path: str) -> str:
 
 
 def _validate_file_path(file_path: str) -> Path:
-    """Validate *file_path* against the project root boundary.
-
-    When ``host_project_root`` is set (e.g. in Docker), paths that look like
-    absolute host paths under that root are mapped to project_root so Cursor
-    and other clients can send host paths without "path denied" errors.
-    """
+    """Validate *file_path* against the project root boundary."""
     from tapps_mcp.security.path_validator import PathValidator
 
     settings = load_settings()
@@ -77,11 +77,7 @@ def _validate_file_path(file_path: str) -> Path:
 
 
 def _record_call(tool_name: str) -> None:
-    """Record a tool call in the session checklist tracker.
-
-    If tapps_mcp.tools.checklist is unavailable (e.g. incomplete binary install),
-    we no-op so other tools still run.
-    """
+    """Record a tool call in the session checklist tracker."""
     try:
         from tapps_mcp.tools.checklist import CallTracker
 
@@ -123,9 +119,11 @@ def _record_execution(
     )
 
 
-# ---------------------------------------------------------------------------
-# Nudge helper — injects next_steps + pipeline_progress into any response
-# ---------------------------------------------------------------------------
+def _get_metrics_hub() -> MetricsHub:
+    """Lazily import and return the global MetricsHub."""
+    from tapps_mcp.metrics.collector import get_metrics_hub
+
+    return get_metrics_hub()
 
 
 def _with_nudges(
@@ -133,10 +131,7 @@ def _with_nudges(
     response: dict[str, Any],
     context: dict[str, Any] | None = None,
 ) -> dict[str, Any]:
-    """Inject ``next_steps`` and ``pipeline_progress`` into a response dict.
-
-    Safe to call on error responses (they are returned unchanged).
-    """
+    """Inject ``next_steps`` and ``pipeline_progress`` into a response dict."""
     if not response.get("success", False):
         return response
     from tapps_mcp.common.nudges import compute_next_steps, compute_pipeline_progress
@@ -152,7 +147,7 @@ def _with_nudges(
 
 
 # ---------------------------------------------------------------------------
-# Tools
+# Core tools (kept in server.py — simple, few dependencies)
 # ---------------------------------------------------------------------------
 
 
@@ -168,44 +163,24 @@ def tapps_server_info() -> dict[str, Any]:
     settings = load_settings()
     installed = detect_installed_tools()
 
-    # Collect subsystem diagnostics (Context7, cache, RAG, knowledge base)
     from tapps_mcp.diagnostics import collect_diagnostics
 
     cache_dir = settings.project_root / ".tapps-mcp-cache"
-    diagnostics = collect_diagnostics(
-        api_key=settings.context7_api_key,
-        cache_dir=cache_dir,
-    )
+    diagnostics = collect_diagnostics(api_key=settings.context7_api_key, cache_dir=cache_dir)
 
-    # Build tool list from the MCP server itself
     available_tools: list[str] = []
     try:
         tool_manager = mcp._tool_manager
         available_tools = list(tool_manager._tools.keys())
     except AttributeError:
-        # Fallback if internal API changes
         available_tools = [
-            "tapps_server_info",
-            "tapps_session_start",
-            "tapps_score_file",
-            "tapps_security_scan",
-            "tapps_quality_gate",
-            "tapps_lookup_docs",
-            "tapps_validate_config",
-            "tapps_validate_changed",
-            "tapps_quick_check",
-            "tapps_consult_expert",
-            "tapps_list_experts",
-            "tapps_checklist",
-            "tapps_project_profile",
-            "tapps_session_notes",
-            "tapps_impact_analysis",
-            "tapps_report",
-            "tapps_init",
-            "tapps_dashboard",
-            "tapps_stats",
-            "tapps_feedback",
-            "tapps_research",
+            "tapps_server_info", "tapps_session_start", "tapps_score_file",
+            "tapps_security_scan", "tapps_quality_gate", "tapps_lookup_docs",
+            "tapps_validate_config", "tapps_validate_changed", "tapps_quick_check",
+            "tapps_consult_expert", "tapps_list_experts", "tapps_checklist",
+            "tapps_project_profile", "tapps_session_notes", "tapps_impact_analysis",
+            "tapps_report", "tapps_init", "tapps_dashboard", "tapps_stats",
+            "tapps_feedback", "tapps_research",
         ]
 
     elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
@@ -214,11 +189,7 @@ def tapps_server_info() -> dict[str, Any]:
     from tapps_mcp.pipeline.models import STAGE_TOOLS, PipelineStage
 
     resp = success_response("tapps_server_info", elapsed_ms, {
-        "server": {
-            "name": "TappsMCP",
-            "version": __version__,
-            "protocol_version": "2025-11-25",
-        },
+        "server": {"name": "TappsMCP", "version": __version__, "protocol_version": "2025-11-25"},
         "configuration": {
             "project_root": str(settings.project_root),
             "quality_preset": settings.quality_preset,
@@ -229,7 +200,7 @@ def tapps_server_info() -> dict[str, Any]:
         "diagnostics": diagnostics.model_dump(),
         "recommended_workflow": (
             "FIRST: Call tapps_session_start() to initialize. "
-            "BEFORE using any library: Call tapps_lookup_docs(). For domain+library questions, pair with tapps_consult_expert(). "
+            "BEFORE using any library: Call tapps_lookup_docs(). "
             "AFTER editing Python files: Call tapps_score_file(quick=True) or tapps_quick_check(). "
             "BEFORE declaring done: Call tapps_validate_changed() or tapps_quality_gate(). "
             "FINAL step: Call tapps_checklist()."
@@ -250,9 +221,7 @@ def tapps_server_info() -> dict[str, Any]:
         "pipeline": {
             "name": "TAPPS Quality Pipeline",
             "stages": [s.value for s in PipelineStage],
-            "current_hint": (
-                "Start with tapps_pipeline_overview prompt, or follow stages in order."
-            ),
+            "current_hint": "Start with tapps_pipeline_overview prompt, or follow stages in order.",
             "stage_tools": {s.value: tools for s, tools in STAGE_TOOLS.items()},
             "handoff_file": "docs/TAPPS_HANDOFF.md",
             "runlog_file": "docs/TAPPS_RUNLOG.md",
@@ -263,131 +232,9 @@ def tapps_server_info() -> dict[str, Any]:
 
 
 @mcp.tool()
-async def tapps_score_file(
-    file_path: str,
-    quick: bool = False,
-    fix: bool = False,
-    mode: str = "auto",
-) -> dict[str, Any]:
-    """REQUIRED after editing any Python file. Scores quality across 7
-    categories (complexity, security, maintainability, test coverage,
-    performance, structure, devex). Skipping means quality issues go
-    undetected and the quality gate will likely fail.
-
-    Use quick=True during edit-lint-fix loops; use full (quick=False) before
-    declaring work complete.
-
-    Args:
-        file_path: Path to the Python file to score.
-        quick: If True, run ruff-only scoring (< 500 ms).
-        fix: If True (requires quick=True), apply ruff auto-fixes first.
-        mode: Execution mode - "subprocess", "direct", or "auto" (default).
-            "direct" uses radon as a library and sync subprocess in thread
-            pool, avoiding async subprocess reliability issues.
-    """
-    start = time.perf_counter_ns()
-    _record_call("tapps_score_file")
-
-    try:
-        resolved = _validate_file_path(file_path)
-    except (ValueError, FileNotFoundError) as exc:
-        return error_response("tapps_score_file", "path_denied", str(exc))
-
-    from tapps_mcp.scoring.scorer import CodeScorer
-
-    scorer = CodeScorer()
-    fixes_applied = 0
-
-    if quick:
-        if fix:
-            from tapps_mcp.tools.ruff import run_ruff_fix
-
-            fixes_applied = run_ruff_fix(str(resolved), cwd=str(resolved.parent))
-
-        result = scorer.score_file_quick(resolved)
-    else:
-        result = await scorer.score_file(resolved, mode=mode)
-
-    elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-
-    # Aggregate suggestions across all categories
-    all_suggestions: list[str] = []
-    for cat in result.categories.values():
-        all_suggestions.extend(cat.suggestions)
-
-    data: dict[str, Any] = {
-        "file_path": result.file_path,
-        "overall_score": round(result.overall_score, 2),
-        "categories": {
-            name: {
-                "score": round(cat.score, 2),
-                "weight": cat.weight,
-                "details": cat.details,
-                "suggestions": cat.suggestions,
-            }
-            for name, cat in result.categories.items()
-        },
-        "suggestions": all_suggestions,
-        "lint_issue_count": len(result.lint_issues),
-        "type_issue_count": len(result.type_issues),
-        "security_issue_count": len(result.security_issues),
-    }
-
-    if quick and fix:
-        data["fixes_applied"] = fixes_applied
-
-    if result.lint_issues:
-        data["lint_issues"] = serialize_issues(result.lint_issues)
-    if result.type_issues:
-        data["type_issues"] = serialize_issues(result.type_issues)
-    if result.security_issues:
-        data["security_issues"] = serialize_issues(result.security_issues)
-    if result.tool_errors:
-        data["tool_errors"] = result.tool_errors
-
-    # Detect uncached library imports (non-critical enhancement)
-    try:
-        from tapps_mcp.knowledge.cache import KBCache
-        from tapps_mcp.knowledge.import_analyzer import (
-            extract_external_imports,
-            find_uncached_libraries,
-        )
-
-        settings_cache = load_settings()
-        cache = KBCache(settings_cache.project_root / ".tapps-mcp-cache")
-        external = extract_external_imports(resolved, settings_cache.project_root)
-        uncached = find_uncached_libraries(external, cache)
-        if uncached:
-            data["uncached_libraries"] = uncached[:10]
-            data["docs_hint"] = (
-                f"Call tapps_lookup_docs for {', '.join(uncached[:5])} "
-                "to avoid hallucinated APIs"
-            )
-    except Exception as exc:  # Never fail scoring for import analysis
-        logger.debug("score_file_import_analysis_skip", file_path=str(resolved), error=str(exc))
-
-    _record_execution(
-        "tapps_score_file",
-        start,
-        file_path=str(resolved),
-        score=round(result.overall_score, 2),
-        degraded=result.degraded,
-    )
-
-    resp = success_response("tapps_score_file", elapsed_ms, data, degraded=result.degraded)
-    return _with_nudges("tapps_score_file", resp, {
-        "security_issue_count": len(result.security_issues),
-    })
-
-
-@mcp.tool()
-def tapps_security_scan(
-    file_path: str,
-    scan_secrets: bool = True,
-) -> dict[str, Any]:
-    """REQUIRED when changes touch security-sensitive code (auth, input handling,
-    secrets, crypto). Runs bandit and secret detection on a Python file.
-    Skipping risks shipping vulnerabilities to production.
+def tapps_security_scan(file_path: str, scan_secrets: bool = True) -> dict[str, Any]:
+    """REQUIRED when changes touch security-sensitive code. Runs bandit and
+    secret detection on a Python file.
 
     Args:
         file_path: Path to the Python file to scan.
@@ -405,123 +252,36 @@ def tapps_security_scan(
 
     settings = load_settings()
     result = run_security_scan(
-        str(resolved),
-        scan_secrets=scan_secrets,
-        cwd=str(settings.project_root),
-        timeout=settings.tool_timeout,
+        str(resolved), scan_secrets=scan_secrets,
+        cwd=str(settings.project_root), timeout=settings.tool_timeout,
     )
 
     elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
     _record_execution(
-        "tapps_security_scan",
-        start,
-        file_path=str(resolved),
-        degraded=not result.bandit_available,
+        "tapps_security_scan", start,
+        file_path=str(resolved), degraded=not result.bandit_available,
     )
 
-    resp = success_response(
-        "tapps_security_scan",
-        elapsed_ms,
-        {
-            "file_path": str(resolved),
-            "passed": result.passed,
-            "total_issues": result.total_issues,
-            "critical_count": result.critical_count,
-            "high_count": result.high_count,
-            "medium_count": result.medium_count,
-            "low_count": result.low_count,
-            "bandit_available": result.bandit_available,
-            "bandit_issues": serialize_issues(result.bandit_issues, limit=30),
-            "secret_findings": serialize_issues(result.secret_findings, limit=30),
-        },
-        degraded=not result.bandit_available,
-    )
+    resp = success_response("tapps_security_scan", elapsed_ms, {
+        "file_path": str(resolved), "passed": result.passed,
+        "total_issues": result.total_issues, "critical_count": result.critical_count,
+        "high_count": result.high_count, "medium_count": result.medium_count,
+        "low_count": result.low_count, "bandit_available": result.bandit_available,
+        "bandit_issues": serialize_issues(result.bandit_issues, limit=30),
+        "secret_findings": serialize_issues(result.secret_findings, limit=30),
+    }, degraded=not result.bandit_available)
     return _with_nudges("tapps_security_scan", resp)
 
 
 @mcp.tool()
-async def tapps_quality_gate(
-    file_path: str,
-    preset: str = "standard",
-) -> dict[str, Any]:
-    """BLOCKING REQUIREMENT before declaring work complete. Runs full scoring
-    then evaluates pass/fail against the quality preset. Work is NOT done
-    until this passes (or the user explicitly accepts the risk).
-
-    Args:
-        file_path: Path to the Python file to evaluate.
-        preset: Quality preset — "standard" (70+), "strict" (80+), or "framework" (75+).
-    """
-    start = time.perf_counter_ns()
-    _record_call("tapps_quality_gate")
-
-    try:
-        resolved = _validate_file_path(file_path)
-    except (ValueError, FileNotFoundError) as exc:
-        return error_response("tapps_quality_gate", "path_denied", str(exc))
-
-    from tapps_mcp.gates.evaluator import evaluate_gate
-    from tapps_mcp.scoring.scorer import CodeScorer
-
-    scorer = CodeScorer()
-    score_result = await scorer.score_file(resolved)
-    gate_result = evaluate_gate(score_result, preset=preset)
-
-    elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-    _record_execution(
-        "tapps_quality_gate",
-        start,
-        file_path=str(resolved),
-        gate_passed=gate_result.passed,
-        score=round(score_result.overall_score, 2),
-        degraded=score_result.degraded,
-    )
-
-    # Collect suggestions for failing categories
-    failing_cats = {f.category for f in gate_result.failures}
-    gate_suggestions: list[str] = []
-    for name, cat in score_result.categories.items():
-        if name in failing_cats and cat.suggestions:
-            gate_suggestions.extend(cat.suggestions)
-
-    gate_data: dict[str, Any] = {
-        "file_path": str(resolved),
-        "passed": gate_result.passed,
-        "preset": gate_result.preset,
-        "overall_score": round(score_result.overall_score, 2),
-        "scores": {k: round(v, 2) for k, v in gate_result.scores.items()},
-        "thresholds": gate_result.thresholds.model_dump(),
-        "failures": [f.model_dump() for f in gate_result.failures],
-        "warnings": gate_result.warnings,
-        "suggestions": gate_suggestions,
-    }
-    if score_result.tool_errors:
-        gate_data["tool_errors"] = score_result.tool_errors
-
-    resp = success_response(
-        "tapps_quality_gate",
-        elapsed_ms,
-        gate_data,
-        degraded=score_result.degraded,
-    )
-    return _with_nudges("tapps_quality_gate", resp, {
-        "gate_passed": gate_result.passed,
-    })
-
-
-@mcp.tool()
 async def tapps_lookup_docs(
-    library: str,
-    topic: str = "overview",
-    mode: str = "code",
+    library: str, topic: str = "overview", mode: str = "code",
 ) -> dict[str, Any]:
     """BLOCKING REQUIREMENT before using any external library API. Returns
-    current docs (Context7 + cache) to prevent hallucinated APIs. Skipping
-    leads to incorrect API usage that must be fixed later. Resolves library
-    names via fuzzy matching; content is safety-filtered before return.
+    current docs (Context7 + cache) to prevent hallucinated APIs.
 
     Args:
-        library: Library name (fuzzy-matched, e.g. "fastapi", "react").
+        library: Library name (fuzzy-matched).
         topic: Specific topic within the library (default "overview").
         mode: "code" for API references, "info" for conceptual guides.
     """
@@ -532,8 +292,7 @@ async def tapps_lookup_docs(
     from tapps_mcp.knowledge.lookup import LookupEngine
 
     settings = load_settings()
-    cache_dir = settings.project_root / ".tapps-mcp-cache"
-    cache = KBCache(cache_dir)
+    cache = KBCache(settings.project_root / ".tapps-mcp-cache")
     engine = LookupEngine(cache, api_key=settings.context7_api_key)
 
     try:
@@ -544,13 +303,10 @@ async def tapps_lookup_docs(
     elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
 
     data: dict[str, Any] = {
-        "library": result.library,
-        "topic": result.topic,
-        "source": result.source,
-        "cache_hit": result.cache_hit,
+        "library": result.library, "topic": result.topic,
+        "source": result.source, "cache_hit": result.cache_hit,
         "response_time_ms": result.response_time_ms,
     }
-
     if result.content is not None:
         data["content"] = result.content
         data["token_estimate"] = len(result.content) // 4
@@ -560,33 +316,24 @@ async def tapps_lookup_docs(
         data["fuzzy_score"] = result.fuzzy_score
 
     _record_execution(
-        "tapps_lookup_docs",
-        start,
+        "tapps_lookup_docs", start,
         status="success" if result.success else "failed",
         error_code="api_key_missing" if (result.error and "API key" in result.error) else None,
     )
 
     response = success_response("tapps_lookup_docs", elapsed_ms, data)
     response["success"] = result.success
-
     if result.error:
         err_code = "api_key_missing" if "API key" in result.error else "lookup_failed"
         response["error"] = {"code": err_code, "message": result.error}
     if result.warning:
         response["warning"] = result.warning
-
     return _with_nudges("tapps_lookup_docs", response)
 
 
 @mcp.tool()
-def tapps_validate_config(
-    file_path: str,
-    config_type: str = "auto",
-) -> dict[str, Any]:
+def tapps_validate_config(file_path: str, config_type: str = "auto") -> dict[str, Any]:
     """REQUIRED when changing Dockerfile, docker-compose, or infra config.
-    Validates against best practices (e.g. non-root user, resource limits).
-    Skipping risks deployment failures. Supports Dockerfile, docker-compose.yml,
-    and WebSocket/MQTT/InfluxDB patterns.
 
     Args:
         file_path: Path to the config file to validate.
@@ -601,7 +348,6 @@ def tapps_validate_config(
         return error_response("tapps_validate_config", "path_denied", str(exc))
 
     content = resolved.read_text(encoding="utf-8")
-
     from tapps_mcp.validators.base import validate_config
 
     explicit_type = None if config_type == "auto" else config_type
@@ -611,12 +357,9 @@ def tapps_validate_config(
     _record_execution("tapps_validate_config", start, file_path=str(resolved))
 
     resp = success_response("tapps_validate_config", elapsed_ms, {
-        "file_path": result.file_path,
-        "config_type": result.config_type,
-        "valid": result.valid,
-        "findings": [f.model_dump() for f in result.findings],
-        "suggestions": result.suggestions,
-        "finding_count": len(result.findings),
+        "file_path": result.file_path, "config_type": result.config_type,
+        "valid": result.valid, "findings": [f.model_dump() for f in result.findings],
+        "suggestions": result.suggestions, "finding_count": len(result.findings),
         "critical_count": sum(1 for f in result.findings if f.severity == "critical"),
         "warning_count": sum(1 for f in result.findings if f.severity == "warning"),
     })
@@ -624,41 +367,29 @@ def tapps_validate_config(
 
 
 @mcp.tool()
-def tapps_consult_expert(
-    question: str,
-    domain: str = "",
-) -> dict[str, Any]:
-    """REQUIRED for domain-specific decisions (security, testing, APIs, DB,
-    architecture). Routes to one of 16 built-in experts and returns
-    RAG-backed guidance with confidence scores. Skipping risks incorrect
-    patterns and missed best practices.
+def tapps_consult_expert(question: str, domain: str = "") -> dict[str, Any]:
+    """REQUIRED for domain-specific decisions. Routes to one of 16 built-in
+    experts and returns RAG-backed guidance with confidence scores.
 
     Args:
         question: The technical question to ask (natural language).
         domain: Optional domain override (e.g. "security", "testing-strategies").
-            When empty, the best domain is auto-detected from the question.
     """
     start = time.perf_counter_ns()
     _record_call("tapps_consult_expert")
 
     from tapps_mcp.experts.engine import consult_expert
 
-    result = consult_expert(
-        question=question,
-        domain=domain or None,
-    )
+    result = consult_expert(question=question, domain=domain or None)
 
     elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
     _record_execution("tapps_consult_expert", start)
 
     resp = success_response("tapps_consult_expert", elapsed_ms, {
-        "domain": result.domain,
-        "expert_id": result.expert_id,
-        "expert_name": result.expert_name,
-        "answer": result.answer,
+        "domain": result.domain, "expert_id": result.expert_id,
+        "expert_name": result.expert_name, "answer": result.answer,
         "confidence": round(result.confidence, 4),
-        "factors": result.factors.model_dump(),
-        "sources": result.sources,
+        "factors": result.factors.model_dump(), "sources": result.sources,
         "chunks_used": result.chunks_used,
         "low_confidence_nudge": result.low_confidence_nudge,
         "suggested_tool": result.suggested_tool,
@@ -673,9 +404,7 @@ def tapps_consult_expert(
 
 @mcp.tool()
 def tapps_list_experts() -> dict[str, Any]:
-    """Call before tapps_consult_expert if unsure which domain fits. Returns
-    the 16 built-in experts with domain, description, and knowledge-base status.
-    """
+    """Returns the 16 built-in experts with domain, description, and knowledge-base status."""
     start = time.perf_counter_ns()
     _record_call("tapps_list_experts")
 
@@ -686,23 +415,17 @@ def tapps_list_experts() -> dict[str, Any]:
     _record_execution("tapps_list_experts", start)
 
     resp = success_response("tapps_list_experts", elapsed_ms, {
-        "expert_count": len(experts),
-        "experts": [e.model_dump() for e in experts],
+        "expert_count": len(experts), "experts": [e.model_dump() for e in experts],
     })
     return _with_nudges("tapps_list_experts", resp)
 
 
 @mcp.tool()
-def tapps_checklist(
-    task_type: str = "review",
-) -> dict[str, Any]:
-    """REQUIRED as the FINAL step before declaring work complete. Reports which
-    tools were called and which required/recommended steps were skipped.
-    Skipping means no verification that the quality process was followed.
+def tapps_checklist(task_type: str = "review") -> dict[str, Any]:
+    """REQUIRED as the FINAL step before declaring work complete.
 
     Args:
         task_type: "feature" | "bugfix" | "refactor" | "security" | "review".
-            Use "review" for general code review.
     """
     start = time.perf_counter_ns()
     _record_call("tapps_checklist")
@@ -719,40 +442,24 @@ def tapps_checklist(
         elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
         _record_execution("tapps_checklist", start)
         fallback_data = {
-            "task_type": task_type,
-            "called": [],
-            "missing_required": [],
-            "missing_recommended": [],
-            "missing_optional": [],
-            "missing_required_hints": [],
-            "missing_recommended_hints": [],
-            "missing_optional_hints": [],
-            "complete": False,
-            "total_calls": 0,
+            "task_type": task_type, "called": [], "missing_required": [],
+            "missing_recommended": [], "missing_optional": [],
+            "missing_required_hints": [], "missing_recommended_hints": [],
+            "missing_optional_hints": [], "complete": False, "total_calls": 0,
             "checklist_unavailable": True,
             "message": (
-                "Module tapps_mcp.tools.checklist is not available (e.g. incomplete "
-                "installation or binary). Other tools work; use tapps_quality_gate and "
-                "tapps_security_scan for verification."
+                "Module tapps_mcp.tools.checklist is not available. "
+                "Use tapps_quality_gate and tapps_security_scan for verification."
             ),
         }
         resp = success_response("tapps_checklist", elapsed_ms, fallback_data)
         return _with_nudges("tapps_checklist", resp, {"complete": False})
 
 
-# ---------------------------------------------------------------------------
-# Epic 4: Project Context & Session Management tools
-# ---------------------------------------------------------------------------
-
-
 @mcp.tool()
-def tapps_project_profile(
-    project_root: str = "",
-) -> dict[str, Any]:
+def tapps_project_profile(project_root: str = "") -> dict[str, Any]:
     """REQUIRED at session start. Detects the project's tech stack, type, CI,
-    Docker, test frameworks, and package managers. Returns quality
-    recommendations tailored to the detected stack. Skipping means
-    recommendations are generic instead of project-specific.
+    Docker, test frameworks, and package managers.
 
     Args:
         project_root: Project root path (default: server's configured root).
@@ -774,7 +481,8 @@ def tapps_project_profile(
     except Exception as exc:
         elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
         _record_execution(
-            "tapps_project_profile", start, status="failed", error_code="profile_failed",
+            "tapps_project_profile", start,
+            status="failed", error_code="profile_failed",
         )
         return error_response("tapps_project_profile", "profile_failed", str(exc))
 
@@ -787,10 +495,8 @@ def tapps_project_profile(
         "project_type": profile.project_type,
         "project_type_confidence": round(profile.project_type_confidence, 2),
         "project_type_reason": profile.project_type_reason,
-        "has_ci": profile.has_ci,
-        "ci_systems": profile.ci_systems,
-        "has_docker": profile.has_docker,
-        "has_tests": profile.has_tests,
+        "has_ci": profile.has_ci, "ci_systems": profile.ci_systems,
+        "has_docker": profile.has_docker, "has_tests": profile.has_tests,
         "test_frameworks": profile.test_frameworks,
         "package_managers": profile.package_managers,
         "quality_recommendations": profile.quality_recommendations,
@@ -798,7 +504,7 @@ def tapps_project_profile(
     return _with_nudges("tapps_project_profile", resp)
 
 
-# Session note store - singleton, created on first use
+# Session note store singleton
 _session_store: SessionNoteStore | None = None
 
 
@@ -814,19 +520,12 @@ def _get_session_store() -> SessionNoteStore:
 
 
 @mcp.tool()
-def tapps_session_notes(
-    action: str,
-    key: str = "",
-    value: str = "",
-) -> dict[str, Any]:
+def tapps_session_notes(action: str, key: str = "", value: str = "") -> dict[str, Any]:
     """Persist notes across the session to avoid losing context.
-
-    Use ``save`` to record a constraint or decision, ``get`` to recall one,
-    ``list`` to see all notes, or ``clear`` to reset.
 
     Args:
         action: "save" | "get" | "list" | "clear".
-        key: Note key (required for save/get, optional for clear).
+        key: Note key (required for save/get).
         value: Note value (required for save).
     """
     start = time.perf_counter_ns()
@@ -838,26 +537,24 @@ def tapps_session_notes(
     if action == "save":
         if not key or not value:
             return error_response(
-                "tapps_session_notes", "missing_params", "save requires key and value",
+                "tapps_session_notes", "missing_params",
+                "save requires key and value",
             )
         note = store.save(key, value)
         data = {"action": "save", "note": note.model_dump()}
-
     elif action == "get":
         if not key:
             return error_response("tapps_session_notes", "missing_params", "get requires key")
         found = store.get(key)
-        note_data = found.model_dump() if found else None
-        data = {"action": "get", "note": note_data, "found": found is not None}
-
+        data = {
+            "action": "get",
+            "note": found.model_dump() if found else None,
+            "found": found is not None,
+        }
     elif action == "list":
-        notes = store.list_all()
-        data = {"action": "list", "notes": [n.model_dump() for n in notes]}
-
+        data = {"action": "list", "notes": [n.model_dump() for n in store.list_all()]}
     elif action == "clear":
-        cleared = store.clear(key or None)
-        data = {"action": "clear", "cleared_count": cleared}
-
+        data = {"action": "clear", "cleared_count": store.clear(key or None)}
     else:
         return error_response(
             "tapps_session_notes", "invalid_action",
@@ -867,20 +564,13 @@ def tapps_session_notes(
     elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
     _record_execution("tapps_session_notes", start)
     data.update(store.metadata())
-
     resp = success_response("tapps_session_notes", elapsed_ms, data)
     return _with_nudges("tapps_session_notes", resp)
 
 
 @mcp.tool()
-def tapps_impact_analysis(
-    file_path: str,
-    change_type: str = "modified",
-) -> dict[str, Any]:
-    """REQUIRED before refactoring or deleting files. Uses AST-based import
-    graph analysis to map the blast radius - direct dependents, transitive
-    dependents, and test files that should be re-run. Skipping risks breaking
-    dependent code.
+def tapps_impact_analysis(file_path: str, change_type: str = "modified") -> dict[str, Any]:
+    """REQUIRED before refactoring or deleting files. Maps the blast radius.
 
     Args:
         file_path: Path to the file being changed.
@@ -903,10 +593,8 @@ def tapps_impact_analysis(
     _record_execution("tapps_impact_analysis", start, file_path=str(resolved))
 
     resp = success_response("tapps_impact_analysis", elapsed_ms, {
-        "changed_file": report.changed_file,
-        "change_type": report.change_type,
-        "severity": report.severity,
-        "total_affected": report.total_affected,
+        "changed_file": report.changed_file, "change_type": report.change_type,
+        "severity": report.severity, "total_affected": report.total_affected,
         "direct_dependents": [d.model_dump() for d in report.direct_dependents],
         "transitive_dependents": [d.model_dump() for d in report.transitive_dependents],
         "test_files": [t.model_dump() for t in report.test_files],
@@ -916,14 +604,8 @@ def tapps_impact_analysis(
 
 
 @mcp.tool()
-async def tapps_report(
-    file_path: str = "",
-    report_format: str = "json",
-) -> dict[str, Any]:
+async def tapps_report(file_path: str = "", report_format: str = "json") -> dict[str, Any]:
     """Generate a quality report combining scoring and gate results.
-
-    When *file_path* is provided, scores that single file.
-    Supports JSON (default), markdown, and HTML output.
 
     Args:
         file_path: Path to a Python file (optional - project-wide if omitted).
@@ -938,7 +620,6 @@ async def tapps_report(
 
     settings = load_settings()
     scorer = CodeScorer()
-
     score_results = []
     gate_results = []
 
@@ -951,14 +632,12 @@ async def tapps_report(
         score_results.append(result)
         gate_results.append(evaluate_gate(result, preset=settings.quality_preset))
     else:
-        # Project-wide: score all .py files in project root (max 20)
         from pathlib import Path as _Path
 
         from tapps_mcp.common.utils import should_skip_path
 
         py_files = sorted(_Path(settings.project_root).rglob("*.py"))
         py_files = [f for f in py_files if not should_skip_path(f)][:20]
-
         for pf in py_files:
             try:
                 result = await scorer.score_file(pf)
@@ -967,707 +646,21 @@ async def tapps_report(
             except (ValueError, OSError, RuntimeError) as e:
                 logger.warning("report_file_skip", file=str(pf), error=str(e))
 
-    report_data = generate_report(
-        score_results,
-        gate_results,
-        report_format=report_format,
-    )
-
+    report_data = generate_report(score_results, gate_results, report_format=report_format)
     elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
     _record_execution("tapps_report", start, file_path=file_path or None)
-
     resp = success_response("tapps_report", elapsed_ms, report_data)
     return _with_nudges("tapps_report", resp)
 
 
 # ---------------------------------------------------------------------------
-# Composite tools — reduce LLM call count
-# ---------------------------------------------------------------------------
-
-
-@mcp.tool()
-def tapps_session_start(
-    project_root: str = "",
-) -> dict[str, Any]:
-    """REQUIRED as the FIRST call in every session. Combines server info
-    and project profile detection in a single call. Skipping means all
-    subsequent tools lack project context and recommendations are generic.
-
-    Args:
-        project_root: Project root path (default: server's configured root).
-    """
-    start = time.perf_counter_ns()
-    _record_call("tapps_session_start")
-
-    # Delegate to existing tools (they record their own calls)
-    info = tapps_server_info()
-    profile = tapps_project_profile(project_root=project_root)
-
-    elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-    _record_execution("tapps_session_start", start)
-
-    data: dict[str, Any] = {
-        "server": info["data"]["server"],
-        "configuration": info["data"]["configuration"],
-        "installed_checkers": info["data"]["installed_checkers"],
-        "diagnostics": info["data"]["diagnostics"],
-        "quick_start": info["data"].get("quick_start", []),
-        "critical_rules": info["data"].get("critical_rules", []),
-        "pipeline": info["data"]["pipeline"],
-    }
-
-    if profile.get("success"):
-        data["project_profile"] = profile["data"]
-    else:
-        data["project_profile"] = None
-        err = profile.get("error")
-        data["project_profile_error"] = (
-            err.get("message", "unknown") if isinstance(err, dict) else (str(err) if err is not None else "unknown")
-        )
-
-    resp = success_response("tapps_session_start", elapsed_ms, data)
-    return _with_nudges("tapps_session_start", resp)
-
-
-@mcp.tool()
-async def tapps_validate_changed(
-    file_paths: str = "",
-    base_ref: str = "HEAD",
-    preset: str = "standard",
-    include_security: bool = True,
-) -> dict[str, Any]:
-    """REQUIRED before declaring work complete on multi-file changes.
-    Detects changed Python files (via git diff) or accepts an explicit
-    comma-separated list. Runs score + quality gate + security scan on each
-    file. Skipping means quality issues in changed files go undetected.
-
-    Args:
-        file_paths: Comma-separated file paths (empty = auto-detect via git diff).
-        base_ref: Git ref to diff against (default: HEAD for unstaged changes).
-        preset: Quality gate preset - "standard", "strict", or "framework".
-        include_security: Whether to run security scan on each file.
-    """
-    start = time.perf_counter_ns()
-    _record_call("tapps_validate_changed")
-
-    from tapps_mcp.gates.evaluator import evaluate_gate
-    from tapps_mcp.scoring.scorer import CodeScorer
-    from tapps_mcp.tools.batch_validator import (
-        MAX_BATCH_FILES,
-        detect_changed_python_files,
-        format_batch_summary,
-    )
-
-    settings = load_settings()
-
-    # Resolve files
-    paths: list[Path] = []
-    if file_paths.strip():
-        for raw_fp in file_paths.split(","):
-            cleaned_fp = raw_fp.strip()
-            if not cleaned_fp:
-                continue
-            try:
-                paths.append(_validate_file_path(cleaned_fp))
-            except (ValueError, FileNotFoundError) as exc:
-                logger.warning("validate_changed_skip", file=cleaned_fp, error=str(exc))
-    else:
-        paths = detect_changed_python_files(settings.project_root, base_ref)
-
-    if not paths:
-        elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-        _record_execution("tapps_validate_changed", start)
-        resp = success_response("tapps_validate_changed", elapsed_ms, {
-            "files_validated": 0,
-            "all_gates_passed": True,
-            "total_security_issues": 0,
-            "results": [],
-            "summary": "No changed Python files found.",
-        })
-        return _with_nudges("tapps_validate_changed", resp)
-
-    # Cap at MAX_BATCH_FILES
-    capped = len(paths) > MAX_BATCH_FILES
-    extra_count = len(paths) - MAX_BATCH_FILES if capped else 0
-    paths = paths[:MAX_BATCH_FILES]
-
-    scorer = CodeScorer()
-    results: list[dict[str, Any]] = []
-
-    for path in paths:
-        file_result: dict[str, Any] = {"file_path": str(path)}
-        try:
-            score = await scorer.score_file(path)
-            file_result["overall_score"] = round(score.overall_score, 2)
-
-            gate = evaluate_gate(score, preset=preset)
-            file_result["gate_passed"] = gate.passed
-            if gate.failures:
-                file_result["gate_failures"] = [f.model_dump() for f in gate.failures]
-
-            if include_security:
-                from tapps_mcp.security.security_scanner import run_security_scan
-
-                sec = run_security_scan(
-                    str(path),
-                    scan_secrets=True,
-                    cwd=str(settings.project_root),
-                    timeout=settings.tool_timeout,
-                )
-                file_result["security_passed"] = sec.passed
-                file_result["security_issues"] = sec.total_issues
-        except Exception as exc:
-            file_result["errors"] = [str(exc)]
-            logger.warning("validate_changed_error", file=str(path), exc_info=True)
-
-        results.append(file_result)
-
-    all_passed = all(r.get("gate_passed", False) for r in results)
-    total_sec = sum(r.get("security_issues", 0) for r in results)
-
-    summary = format_batch_summary(results)
-    if capped:
-        summary += f" ({extra_count} additional files not validated - cap {MAX_BATCH_FILES})"
-
-    elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-    _record_execution(
-        "tapps_validate_changed", start,
-        gate_passed=all_passed,
-    )
-
-    resp = success_response("tapps_validate_changed", elapsed_ms, {
-        "files_validated": len(results),
-        "all_gates_passed": all_passed,
-        "total_security_issues": total_sec,
-        "results": results,
-        "summary": summary,
-    })
-    return _with_nudges("tapps_validate_changed", resp)
-
-
-@mcp.tool()
-async def tapps_quick_check(
-    file_path: str,
-    preset: str = "standard",
-) -> dict[str, Any]:
-    """REQUIRED at minimum after editing any Python file. Runs quick
-    score + quality gate + basic security check in one fast call. For
-    thorough validation, use tapps_validate_changed or individual tools.
-    Skipping means quality regressions go unnoticed.
-
-    Args:
-        file_path: Path to the Python file to check.
-        preset: Quality gate preset - "standard", "strict", or "framework".
-    """
-    start = time.perf_counter_ns()
-    _record_call("tapps_quick_check")
-
-    try:
-        resolved = _validate_file_path(file_path)
-    except (ValueError, FileNotFoundError) as exc:
-        return error_response("tapps_quick_check", "path_denied", str(exc))
-
-    from tapps_mcp.gates.evaluator import evaluate_gate
-    from tapps_mcp.scoring.scorer import CodeScorer
-    from tapps_mcp.security.security_scanner import run_security_scan
-
-    settings = load_settings()
-    scorer = CodeScorer()
-
-    # Quick score (ruff-only, fast)
-    score_result = scorer.score_file_quick(resolved)
-
-    # Gate evaluation
-    gate_result = evaluate_gate(score_result, preset=preset)
-
-    # Security scan
-    sec_result = run_security_scan(
-        str(resolved),
-        scan_secrets=True,
-        cwd=str(settings.project_root),
-        timeout=settings.tool_timeout,
-    )
-
-    elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-    _record_execution(
-        "tapps_quick_check",
-        start,
-        file_path=str(resolved),
-        gate_passed=gate_result.passed,
-        score=round(score_result.overall_score, 2),
-    )
-
-    data: dict[str, Any] = {
-        "file_path": str(resolved),
-        "overall_score": round(score_result.overall_score, 2),
-        "gate_passed": gate_result.passed,
-        "gate_preset": preset,
-        "security_passed": sec_result.passed,
-        "lint_issue_count": len(score_result.lint_issues),
-        "security_issue_count": sec_result.total_issues,
-    }
-
-    if gate_result.failures:
-        data["gate_failures"] = [f.model_dump() for f in gate_result.failures]
-    if score_result.lint_issues:
-        data["lint_issues"] = serialize_issues(score_result.lint_issues)
-    if sec_result.total_issues > 0:
-        data["security_issues"] = serialize_issues(
-            sec_result.bandit_issues + sec_result.secret_findings, limit=10,
-        )
-
-    # Aggregate suggestions
-    suggestions: list[str] = []
-    for cat in score_result.categories.values():
-        suggestions.extend(cat.suggestions)
-    if suggestions:
-        data["suggestions"] = suggestions
-
-    resp = success_response(
-        "tapps_quick_check", elapsed_ms, data,
-        degraded=not sec_result.bandit_available,
-    )
-    return _with_nudges("tapps_quick_check", resp)
-
-
-# ---------------------------------------------------------------------------
-# Epic 8: Pipeline Orchestration — MCP Prompts & Bootstrap Tool
-# ---------------------------------------------------------------------------
-
-
-@mcp.prompt()
-def tapps_pipeline(stage: str = "discover") -> str:
-    """TAPPS quality pipeline - structured 5-stage workflow.
-
-    Get instructions for a specific pipeline stage. Stages run in order:
-    discover -> research -> develop -> validate -> verify.
-
-    Args:
-        stage: Pipeline stage to get instructions for.
-               One of: discover, research, develop, validate, verify.
-    """
-    from tapps_mcp.prompts.prompt_loader import load_stage_prompt
-
-    return load_stage_prompt(stage)
-
-
-@mcp.prompt()
-def tapps_pipeline_overview() -> str:
-    """Get a summary of the full TAPPS 5-stage quality pipeline.
-
-    Returns stage names, tool assignments, flow diagram, and
-    handoff file format. Use this to understand the full pipeline
-    before starting.
-    """
-    from tapps_mcp.prompts.prompt_loader import load_overview
-
-    return load_overview()
-
-
-@mcp.tool()
-def tapps_init(
-    create_handoff: bool = True,
-    create_runlog: bool = True,
-    create_agents_md: bool = True,
-    create_tech_stack_md: bool = True,
-    platform: str = "",
-    verify_server: bool = True,
-    install_missing_checkers: bool = False,
-    warm_cache_from_tech_stack: bool = True,
-    warm_expert_rag_from_tech_stack: bool = True,
-    overwrite_platform_rules: bool = False,
-    overwrite_agents_md: bool = False,
-) -> dict[str, Any]:
-    """Bootstrap TAPPS pipeline in the current project.
-
-    Verifies server info and optionally installs missing checkers (ruff, mypy,
-    bandit, radon). Creates handoff, runlog, AGENTS.md, and TECH_STACK.md.
-    Optionally warms the Context7 cache from the detected tech stack.
-    Optionally generates platform-specific rule files for Claude Code or Cursor.
-
-    Call once per project to set up the pipeline workflow.
-
-    Args:
-        create_handoff: Create docs/TAPPS_HANDOFF.md template.
-        create_runlog: Create docs/TAPPS_RUNLOG.md template.
-        create_agents_md: Create AGENTS.md with AI assistant workflow (if missing).
-        create_tech_stack_md: Create or update TECH_STACK.md from project profile.
-        platform: Generate platform rules. One of: "claude", "cursor", "".
-        verify_server: Verify server info and installed checkers.
-        install_missing_checkers: Attempt to pip-install missing checkers (opt-in).
-        warm_cache_from_tech_stack: Pre-fetch docs for tech stack libraries into cache.
-        warm_expert_rag_from_tech_stack: Pre-build expert RAG indices for relevant domains.
-        overwrite_platform_rules: When ``True``, refresh platform rule files even if
-            they already exist (useful when templates are upgraded).
-        overwrite_agents_md: When ``True``, replace AGENTS.md entirely with the latest
-            template. When ``False`` (default), validate and smart-merge missing
-            sections/tools.
-    """
-    start = time.perf_counter_ns()
-    _record_call("tapps_init")
-
-    from tapps_mcp.pipeline.init import bootstrap_pipeline
-
-    settings = load_settings()
-    result = bootstrap_pipeline(
-        settings.project_root,
-        create_handoff=create_handoff,
-        create_runlog=create_runlog,
-        create_agents_md=create_agents_md,
-        create_tech_stack_md=create_tech_stack_md,
-        platform=platform,
-        verify_server=verify_server,
-        install_missing_checkers=install_missing_checkers,
-        warm_cache_from_tech_stack=warm_cache_from_tech_stack,
-        warm_expert_rag_from_tech_stack=warm_expert_rag_from_tech_stack,
-        overwrite_platform_rules=overwrite_platform_rules,
-        overwrite_agents_md=overwrite_agents_md,
-    )
-
-    elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-    _record_execution(
-        "tapps_init",
-        start,
-        status="success" if not result["errors"] else "failed",
-    )
-
-    resp = success_response("tapps_init", elapsed_ms, result)
-    resp["success"] = not result["errors"]
-    return _with_nudges("tapps_init", resp)
-
-
-# ---------------------------------------------------------------------------
-# Epic 7: Metrics, Observability & Dashboard tools
-# ---------------------------------------------------------------------------
-
-
-def _get_metrics_hub() -> MetricsHub:
-    """Lazily import and return the global MetricsHub."""
-    from tapps_mcp.metrics.collector import get_metrics_hub
-
-    return get_metrics_hub()
-
-
-_PERIOD_DAYS: dict[str, int] = {"1d": 1, "7d": 7, "30d": 30}
-
-
-def _session_stats(
-    hub: MetricsHub, tool_name: str | None,
-) -> tuple[Any, list[dict[str, Any]]]:
-    """Compute stats from in-memory session data."""
-    from tapps_mcp.metrics.execution_metrics import ToolCallMetricsCollector
-
-    recent = hub.execution.get_recent(limit=100)
-    summary = ToolCallMetricsCollector._compute_summary(recent)
-
-    by_tool: dict[str, list[Any]] = {}
-    for m in recent:
-        by_tool.setdefault(m.tool_name, []).append(m)
-
-    breakdowns = []
-    for tname, tmetrics in sorted(by_tool.items()):
-        if tool_name and tname != tool_name:
-            continue
-        ts = ToolCallMetricsCollector._compute_summary(tmetrics)
-        breakdowns.append({
-            "tool_name": tname,
-            "call_count": ts.total_calls,
-            "success_rate": ts.success_rate,
-            "avg_duration_ms": ts.avg_duration_ms,
-            "p95_duration_ms": ts.p95_duration_ms,
-        })
-    return summary, breakdowns
-
-
-def _period_stats(
-    hub: MetricsHub, tool_name: str | None, period: str,
-) -> tuple[Any, list[dict[str, Any]]]:
-    """Compute stats from persisted data for a given time period."""
-    from datetime import UTC, datetime, timedelta
-
-    since: datetime | None = None
-    days = _PERIOD_DAYS.get(period)
-    if days is not None:
-        since = datetime.now(tz=UTC) - timedelta(days=days)
-
-    summary = hub.execution.get_summary(since=since)
-    raw = hub.execution.get_summary_by_tool(since=since)
-    breakdowns = [
-        {
-            "tool_name": b.tool_name,
-            "call_count": b.call_count,
-            "success_rate": b.success_rate,
-            "avg_duration_ms": b.avg_duration_ms,
-            "p95_duration_ms": b.p95_duration_ms,
-        }
-        for b in raw
-        if not tool_name or b.tool_name == tool_name
-    ]
-    return summary, breakdowns
-
-
-@mcp.tool()
-async def tapps_dashboard(
-    output_format: str = "json",
-    time_range: str = "7d",
-    sections: list[str] | None = None,
-) -> dict[str, Any]:
-    """Generate a comprehensive metrics dashboard.
-
-    Call this to review how well TappsMCP is performing - scoring accuracy,
-    gate pass rates, expert effectiveness, cache performance, quality trends,
-    and alerts.
-
-    Args:
-        output_format: Output format - "json" (default), "markdown", "html", or "otel".
-        time_range: Time range - "1d", "7d", "30d", "90d".
-        sections: Specific sections to include (default: all).
-            Options: summary, tool_metrics, scoring_trends, expert_metrics,
-            cache_metrics, quality_distribution, alerts, business_metrics,
-            recommendations.
-    """
-    start = time.perf_counter_ns()
-    _record_call("tapps_dashboard")
-
-    hub = _get_metrics_hub()
-
-    if output_format == "otel":
-        from tapps_mcp.metrics.otel_export import export_otel_trace
-
-        recent = hub.execution.get_recent(limit=100)
-        otel_data = export_otel_trace(recent)
-        elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-        _record_execution("tapps_dashboard", start)
-        resp = success_response("tapps_dashboard", elapsed_ms, otel_data)
-        return _with_nudges("tapps_dashboard", resp)
-
-    dashboard = hub.get_dashboard_generator()
-
-    if output_format == "json":
-        data = dashboard.generate_json_dashboard(sections=sections)
-    elif output_format == "markdown":
-        content = dashboard.generate_markdown_dashboard(sections=sections)
-        data = {"format": "markdown", "content": content}
-    elif output_format == "html":
-        content = dashboard.generate_html_dashboard(sections=sections)
-        path = dashboard.save_dashboard(fmt="html", sections=sections)
-        data = {"format": "html", "content": content, "saved_to": str(path)}
-    else:
-        data = dashboard.generate_json_dashboard(sections=sections)
-
-    elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-    _record_execution("tapps_dashboard", start)
-    resp = success_response("tapps_dashboard", elapsed_ms, data)
-    return _with_nudges("tapps_dashboard", resp)
-
-
-@mcp.tool()
-def tapps_stats(
-    tool_name: str | None = None,
-    period: str = "session",
-) -> dict[str, Any]:
-    """Return usage statistics for TappsMCP tools.
-
-    Shows call counts, success rates, average durations, cache hit rates,
-    and gate pass rates.
-
-    Args:
-        tool_name: Filter stats to a specific tool (optional).
-        period: Stats period - "session", "1d", "7d", "30d", "all".
-    """
-    start = time.perf_counter_ns()
-    _record_call("tapps_stats")
-
-    hub = _get_metrics_hub()
-
-    if period == "session":
-        summary, tool_breakdowns = _session_stats(hub, tool_name)
-    else:
-        summary, tool_breakdowns = _period_stats(hub, tool_name, period)
-
-    elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-    _record_execution("tapps_stats", start)
-
-    resp = success_response("tapps_stats", elapsed_ms, {
-        "period": period,
-        "total_calls": summary.total_calls,
-        "success_rate": summary.success_rate,
-        "avg_duration_ms": summary.avg_duration_ms,
-        "p95_duration_ms": summary.p95_duration_ms,
-        "gate_pass_rate": summary.gate_pass_rate,
-        "avg_score": summary.avg_score,
-        "tools": tool_breakdowns,
-    })
-    return _with_nudges("tapps_stats", resp)
-
-
-@mcp.tool()
-def tapps_feedback(
-    tool_name: str,
-    helpful: bool,
-    context: str | None = None,
-) -> dict[str, Any]:
-    """Report whether a tool's output was helpful.
-
-    This feedback improves TappsMCP's adaptive scoring and expert weights
-    over time.
-
-    Args:
-        tool_name: Which tool to provide feedback on.
-        helpful: Was the output helpful?
-        context: Additional context about why it was or wasn't helpful.
-    """
-    start = time.perf_counter_ns()
-    _record_call("tapps_feedback")
-
-    from tapps_mcp.metrics.feedback import FeedbackTracker
-
-    hub = _get_metrics_hub()
-    tracker = FeedbackTracker(hub.metrics_dir)
-
-    tracker.record(
-        tool_name=tool_name,
-        helpful=helpful,
-        context=context or "",
-        session_id=hub.session_id,
-    )
-
-    stats = tracker.get_statistics(tool_name=tool_name)
-    overall_stats = tracker.get_statistics()
-
-    elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-    _record_execution("tapps_feedback", start)
-
-    resp = success_response("tapps_feedback", elapsed_ms, {
-        "recorded": True,
-        "tool_name": tool_name,
-        "helpful": helpful,
-        "tool_stats": stats,
-        "overall_stats": overall_stats,
-    })
-    return _with_nudges("tapps_feedback", resp)
-
-
-@mcp.tool()
-def tapps_research(
-    question: str,
-    domain: str = "",
-    library: str = "",
-    topic: str = "",
-) -> dict[str, Any]:
-    """Combined expert consultation + documentation lookup in one call.
-
-    Consults the domain expert first, then automatically supplements with
-    Context7 documentation when expert RAG has no results or confidence is
-    low.  Saves a round-trip compared to calling tapps_consult_expert and
-    tapps_lookup_docs separately.
-
-    Args:
-        question: The technical question to research (natural language).
-        domain: Optional domain override for expert routing.
-        library: Optional library name for docs lookup (auto-inferred when empty).
-        topic: Optional topic for docs lookup (auto-inferred when empty).
-    """
-    start = time.perf_counter_ns()
-    _record_call("tapps_research")
-
-    from tapps_mcp.experts.engine import consult_expert
-
-    result = consult_expert(
-        question=question,
-        domain=domain or None,
-    )
-
-    # Determine whether we need a docs supplement.
-    needs_docs = result.chunks_used == 0 or result.confidence < 0.5
-
-    docs_content: str | None = None
-    docs_source: str | None = None
-    docs_library: str | None = None
-    docs_topic: str | None = None
-    docs_error: str | None = None
-
-    if needs_docs:
-        # Resolve library/topic: prefer explicit args, fall back to expert hints.
-        lookup_library = library or result.suggested_library or "python"
-        lookup_topic = topic or result.suggested_topic or "overview"
-
-        try:
-            import asyncio
-
-            from tapps_mcp.knowledge.cache import KBCache
-            from tapps_mcp.knowledge.lookup import LookupEngine
-
-            settings = load_settings()
-            cache = KBCache(settings.project_root / ".tapps-mcp-cache")
-
-            async def _run() -> tuple[bool, str | None, str | None]:
-                engine = LookupEngine(cache, api_key=settings.context7_api_key)
-                try:
-                    lr = await engine.lookup(library=lookup_library, topic=lookup_topic, mode="code")
-                finally:
-                    await engine.close()
-                return lr.success, lr.content, lr.source
-
-            ok, content, source = asyncio.run(_run())
-            if ok and content:
-                max_chars = settings.expert_fallback_max_chars
-                docs_content = content[:max_chars]
-                docs_source = source
-                docs_library = lookup_library
-                docs_topic = lookup_topic
-            else:
-                docs_error = "Docs lookup returned no content."
-        except Exception as exc:
-            docs_error = str(exc)
-
-    # Build merged answer.
-    answer = result.answer
-    if docs_content:
-        answer = (
-            f"{answer}\n\n---\n\n"
-            f"### Documentation reference (auto-attached)\n\n"
-            f"Library: `{docs_library}` | Topic: `{docs_topic}` | Source: {docs_source}\n\n"
-            f"{docs_content}"
-        )
-
-    elapsed_ms = (time.perf_counter_ns() - start) // 1_000_000
-    _record_execution("tapps_research", start)
-
-    resp = success_response("tapps_research", elapsed_ms, {
-        "domain": result.domain,
-        "expert_id": result.expert_id,
-        "expert_name": result.expert_name,
-        "answer": answer,
-        "confidence": round(result.confidence, 4),
-        "factors": result.factors.model_dump(),
-        "sources": result.sources,
-        "chunks_used": result.chunks_used,
-        "docs_supplemented": docs_content is not None,
-        "docs_library": docs_library,
-        "docs_topic": docs_topic,
-        "docs_error": docs_error,
-        "suggested_tool": result.suggested_tool,
-        "suggested_library": result.suggested_library,
-        "suggested_topic": result.suggested_topic,
-        "fallback_used": result.fallback_used,
-        "fallback_library": result.fallback_library,
-        "fallback_topic": result.fallback_topic,
-    })
-    return _with_nudges("tapps_research", resp)
-
-
-# ---------------------------------------------------------------------------
-# MCP Resources -- browsable knowledge and configuration
+# MCP Resources
 # ---------------------------------------------------------------------------
 
 
 @mcp.resource("tapps://knowledge/{domain}/{topic}")
 def get_knowledge_resource(domain: str, topic: str) -> str:
-    """Retrieve expert knowledge for a domain and topic.
-
-    Browse the 119 knowledge files across 16 expert domains.
-    """
+    """Retrieve expert knowledge for a domain and topic."""
     import re
     from pathlib import Path
 
@@ -1677,25 +670,22 @@ def get_knowledge_resource(domain: str, topic: str) -> str:
         valid = ", ".join(sorted(ExpertRegistry.TECHNICAL_DOMAINS))
         return f"Unknown domain: {domain}. Valid domains: {valid}"
 
-    # Sanitise topic to prevent path traversal (alphanumeric, hyphens, underscores only)
     if not re.match(r"^[a-zA-Z0-9_-]+$", topic):
         return f"Invalid topic name: '{topic}'. Use only alphanumeric, hyphens, underscores."
 
     knowledge_dir = Path(__file__).parent / "experts" / "knowledge" / domain
     topic_file = knowledge_dir / f"{topic}.md"
 
-    # Verify resolved path stays within the knowledge directory
     try:
         topic_file.resolve().relative_to(knowledge_dir.resolve())
     except ValueError:
         return f"Invalid topic path: '{topic}'."
 
     if not topic_file.exists():
-        # List available topics
         if knowledge_dir.exists():
             available = [f.stem for f in knowledge_dir.glob("*.md")]
-            avail_str = ", ".join(sorted(available))
-            return f"Topic '{topic}' not found in domain '{domain}'. Available: {avail_str}"
+            avail = ", ".join(sorted(available))
+            return f"Topic '{topic}' not found in domain '{domain}'. Available: {avail}"
         return f"No knowledge directory for domain '{domain}'."
 
     return topic_file.read_text(encoding="utf-8")
@@ -1708,7 +698,6 @@ def list_knowledge_domains() -> str:
 
     knowledge_base = Path(__file__).parent / "experts" / "knowledge"
     lines = ["# TappsMCP Knowledge Domains\n"]
-
     for domain_dir in sorted(knowledge_base.iterdir()):
         if not domain_dir.is_dir():
             continue
@@ -1717,7 +706,6 @@ def list_knowledge_domains() -> str:
         lines.append(f"Topics ({len(topics)}):")
         for t in topics:
             lines.append(f"  - {t}")
-
     return "\n".join(lines)
 
 
@@ -1732,7 +720,6 @@ def get_quality_presets() -> str:
         for key, value in thresholds.items():
             lines.append(f"  {key}: {value}")
         lines.append("")
-
     return "\n".join(lines)
 
 
@@ -1741,26 +728,42 @@ def get_scoring_weights() -> str:
     """Get current scoring category weights."""
     settings = load_settings()
     w = settings.scoring_weights
-
-    lines = ["# Scoring Weights\n"]
     from tapps_mcp.config.settings import ScoringWeights
 
+    lines = ["# Scoring Weights\n"]
     for field_name in ScoringWeights.model_fields:
         lines.append(f"  {field_name}: {getattr(w, field_name)}")
-
     return "\n".join(lines)
 
 
 # ---------------------------------------------------------------------------
-# MCP Prompts -- workflow guidance
+# MCP Prompts
 # ---------------------------------------------------------------------------
+
+
+@mcp.prompt()
+def tapps_pipeline(stage: str = "discover") -> str:
+    """TAPPS quality pipeline - structured 5-stage workflow.
+
+    Args:
+        stage: Pipeline stage to get instructions for.
+    """
+    from tapps_mcp.prompts.prompt_loader import load_stage_prompt
+
+    return load_stage_prompt(stage)
+
+
+@mcp.prompt()
+def tapps_pipeline_overview() -> str:
+    """Get a summary of the full TAPPS 5-stage quality pipeline."""
+    from tapps_mcp.prompts.prompt_loader import load_overview
+
+    return load_overview()
 
 
 @mcp.prompt()
 def tapps_workflow(task_type: str = "general") -> str:
     """Generate the TappsMCP workflow prompt for a specific task type.
-
-    Provides tool call order and recommendations based on the task type.
 
     Args:
         task_type: One of: general, feature, bugfix, refactor, security, review.
@@ -1768,63 +771,73 @@ def tapps_workflow(task_type: str = "general") -> str:
     workflows = {
         "general": (
             "TappsMCP Workflow - General\n\n"
-            "1. tapps_server_info - discover capabilities\n"
-            "2. tapps_project_profile - understand the project\n"
-            "3. tapps_score_file(quick=True) - quick scoring during edits\n"
-            "4. tapps_score_file - full scoring before completion\n"
-            "5. tapps_quality_gate - verify quality bar\n"
-            "6. tapps_checklist(task_type='review') - verify completeness"
+            "1. tapps_server_info\n2. tapps_project_profile\n"
+            "3. tapps_score_file(quick=True)\n4. tapps_score_file\n"
+            "5. tapps_quality_gate\n6. tapps_checklist(task_type='review')"
         ),
         "feature": (
             "TappsMCP Workflow - New Feature\n\n"
-            "1. tapps_server_info - discover capabilities\n"
-            "2. tapps_project_profile - understand the project\n"
-            "3. tapps_lookup_docs - check library APIs before coding\n"
-            "4. tapps_consult_expert - get domain guidance\n"
-            "5. tapps_score_file(quick=True) - quick scoring during edits\n"
-            "6. tapps_score_file - full scoring\n"
-            "7. tapps_security_scan - check for vulnerabilities\n"
-            "8. tapps_quality_gate - verify quality bar\n"
-            "9. tapps_checklist(task_type='feature') - verify completeness"
+            "1. tapps_server_info\n2. tapps_project_profile\n"
+            "3. tapps_lookup_docs\n4. tapps_consult_expert\n"
+            "5. tapps_score_file(quick=True)\n6. tapps_score_file\n"
+            "7. tapps_security_scan\n8. tapps_quality_gate\n"
+            "9. tapps_checklist(task_type='feature')"
         ),
         "bugfix": (
             "TappsMCP Workflow - Bug Fix\n\n"
-            "1. tapps_server_info - discover capabilities\n"
-            "2. tapps_impact_analysis - understand blast radius\n"
-            "3. tapps_score_file(quick=True) - quick scoring during fix\n"
-            "4. tapps_score_file - full scoring after fix\n"
-            "5. tapps_quality_gate - verify quality bar\n"
-            "6. tapps_checklist(task_type='bugfix') - verify completeness"
+            "1. tapps_server_info\n2. tapps_impact_analysis\n"
+            "3. tapps_score_file(quick=True)\n4. tapps_score_file\n"
+            "5. tapps_quality_gate\n6. tapps_checklist(task_type='bugfix')"
         ),
         "refactor": (
             "TappsMCP Workflow - Refactoring\n\n"
-            "1. tapps_server_info - discover capabilities\n"
-            "2. tapps_impact_analysis - understand dependencies\n"
-            "3. tapps_consult_expert(domain='software-architecture') - get guidance\n"
-            "4. tapps_score_file(quick=True) - quick scoring during refactor\n"
-            "5. tapps_score_file - full scoring\n"
-            "6. tapps_quality_gate - verify quality bar\n"
-            "7. tapps_checklist(task_type='refactor') - verify completeness"
+            "1. tapps_server_info\n2. tapps_impact_analysis\n"
+            "3. tapps_consult_expert(domain='software-architecture')\n"
+            "4. tapps_score_file(quick=True)\n5. tapps_score_file\n"
+            "6. tapps_quality_gate\n7. tapps_checklist(task_type='refactor')"
         ),
         "security": (
             "TappsMCP Workflow - Security Review\n\n"
-            "1. tapps_server_info - discover capabilities\n"
-            "2. tapps_security_scan - comprehensive security scan\n"
-            "3. tapps_consult_expert(domain='security') - get security guidance\n"
-            "4. tapps_score_file - full scoring with security focus\n"
-            "5. tapps_quality_gate(preset='strict') - strict quality bar\n"
-            "6. tapps_checklist(task_type='security') - verify completeness"
+            "1. tapps_server_info\n2. tapps_security_scan\n"
+            "3. tapps_consult_expert(domain='security')\n"
+            "4. tapps_score_file\n5. tapps_quality_gate(preset='strict')\n"
+            "6. tapps_checklist(task_type='security')"
         ),
         "review": (
             "TappsMCP Workflow - Code Review\n\n"
-            "1. tapps_server_info - discover capabilities\n"
-            "2. tapps_score_file - full scoring\n"
-            "3. tapps_security_scan - security check\n"
-            "4. tapps_quality_gate - verify quality bar\n"
-            "5. tapps_checklist(task_type='review') - verify completeness"
+            "1. tapps_server_info\n2. tapps_score_file\n"
+            "3. tapps_security_scan\n4. tapps_quality_gate\n"
+            "5. tapps_checklist(task_type='review')"
         ),
     }
     return workflows.get(task_type, workflows["general"])
+
+
+# ---------------------------------------------------------------------------
+# Register tools from extracted modules & re-export for backward compatibility
+# ---------------------------------------------------------------------------
+
+from tapps_mcp import (  # noqa: E402
+    server_metrics_tools,
+    server_pipeline_tools,
+    server_scoring_tools,
+)
+
+server_scoring_tools.register(mcp)
+server_pipeline_tools.register(mcp)
+server_metrics_tools.register(mcp)
+
+# Re-export so ``from tapps_mcp.server import tapps_X`` keeps working
+tapps_score_file = server_scoring_tools.tapps_score_file
+tapps_quality_gate = server_scoring_tools.tapps_quality_gate
+tapps_quick_check = server_scoring_tools.tapps_quick_check
+tapps_validate_changed = server_pipeline_tools.tapps_validate_changed
+tapps_session_start = server_pipeline_tools.tapps_session_start
+tapps_init = server_pipeline_tools.tapps_init
+tapps_dashboard = server_metrics_tools.tapps_dashboard
+tapps_stats = server_metrics_tools.tapps_stats
+tapps_feedback = server_metrics_tools.tapps_feedback
+tapps_research = server_metrics_tools.tapps_research
 
 
 # ---------------------------------------------------------------------------
@@ -1837,13 +850,7 @@ def run_server(
     host: str = "127.0.0.1",
     port: int = 8000,
 ) -> None:
-    """Start the TappsMCP MCP server.
-
-    Args:
-        transport: ``"stdio"`` for local MCP hosts, ``"http"`` for remote.
-        host: Bind address for HTTP transport.
-        port: Bind port for HTTP transport.
-    """
+    """Start the TappsMCP MCP server."""
     settings = load_settings()
     setup_logging(level=settings.log_level, json_output=settings.log_json)
 
@@ -1858,7 +865,6 @@ def run_server(
     if transport == "stdio":
         mcp.run(transport="stdio")
     elif transport == "http":
-        # Streamable HTTP via uvicorn; wrap so GET / returns a simple "running" page
         import uvicorn
         from starlette.requests import Request  # noqa: TC002
         from starlette.responses import HTMLResponse
@@ -1874,9 +880,6 @@ def run_server(
                 status_code=200,
             )
 
-        # Add root route directly to the MCP app so its lifespan
-        # (task group init) runs properly — wrapping in a parent
-        # Starlette breaks the lifespan chain.
         mcp_app.routes.insert(0, Route("/", _root))
         uvicorn.run(mcp_app, host=host, port=port)
     else:
